RendererDRMPRIME: extract VideoLayerBridgeDRMPRIME
