+++ xbmc/cores/VideoPlayer/VideoRenderers/HwDecRender/RendererDRMPRIME.cpp
@@ -8,16 +8,19 @@
 
 #include "RendererDRMPRIME.h"
 
+#include "cores/VideoPlayer/DVDCodecs/Video/DVDVideoCodecDRMPRIME.h"
+#include "cores/VideoPlayer/VideoRenderers/HwDecRender/VideoLayerBridgeDRMPRIME.h"
 #include "cores/VideoPlayer/VideoRenderers/RenderCapture.h"
 #include "cores/VideoPlayer/VideoRenderers/RenderFactory.h"
 #include "cores/VideoPlayer/VideoRenderers/RenderFlags.h"
-#include "ServiceBroker.h"
-#include "settings/DisplaySettings.h"
 #include "settings/lib/Setting.h"
+#include "settings/DisplaySettings.h"
 #include "settings/Settings.h"
 #include "utils/log.h"
 #include "windowing/gbm/DRMAtomic.h"
+#include "windowing/gbm/WinSystemGbm.h"
 #include "windowing/GraphicContext.h"
+#include "ServiceBroker.h"
 
 const std::string SETTING_VIDEOPLAYER_USEPRIMERENDERER = "videoplayer.useprimerenderer";
 
@@ -31,7 +34,7 @@ CBaseRenderer* CRendererDRMPRIME::Create(CVideoBuffer* buffer)
   if (buffer && dynamic_cast<CVideoBufferDRMPRIME*>(buffer) &&
       CServiceBroker::GetSettings()->GetInt(SETTING_VIDEOPLAYER_USEPRIMERENDERER) == 0)
   {
-    CWinSystemGbmEGLContext* winSystem = dynamic_cast<CWinSystemGbmEGLContext*>(CServiceBroker::GetWinSystem());
+    CWinSystemGbm* winSystem = dynamic_cast<CWinSystemGbm*>(CServiceBroker::GetWinSystem());
     if (winSystem && winSystem->GetDrm()->GetPrimaryPlane()->plane &&
         std::dynamic_pointer_cast<CDRMAtomic>(winSystem->GetDrm()))
       return new CRendererDRMPRIME();
@@ -42,7 +45,7 @@ CBaseRenderer* CRendererDRMPRIME::Create(CVideoBuffer* buffer)
 
 void CRendererDRMPRIME::Register()
 {
-  CWinSystemGbmEGLContext* winSystem = dynamic_cast<CWinSystemGbmEGLContext*>(CServiceBroker::GetWinSystem());
+  CWinSystemGbm* winSystem = dynamic_cast<CWinSystemGbm*>(CServiceBroker::GetWinSystem());
   if (winSystem && winSystem->GetDrm()->GetPrimaryPlane()->plane &&
       std::dynamic_pointer_cast<CDRMAtomic>(winSystem->GetDrm()))
   {
@@ -163,7 +166,7 @@ void CRendererDRMPRIME::RenderUpdate(int index, int index2, bool clear, unsigned
 
   if (!m_videoLayerBridge)
   {
-    CWinSystemGbmEGLContext* winSystem = static_cast<CWinSystemGbmEGLContext*>(CServiceBroker::GetWinSystem());
+    CWinSystemGbm* winSystem = static_cast<CWinSystemGbm*>(CServiceBroker::GetWinSystem());
     m_videoLayerBridge = std::dynamic_pointer_cast<CVideoLayerBridgeDRMPRIME>(winSystem->GetVideoLayerBridge());
     if (!m_videoLayerBridge)
       m_videoLayerBridge = std::make_shared<CVideoLayerBridgeDRMPRIME>(winSystem->GetDrm());
@@ -206,144 +209,4 @@ bool CRendererDRMPRIME::Supports(ERENDERFEATURE feature)
 bool CRendererDRMPRIME::Supports(ESCALINGMETHOD method)
 {
   return false;
-}
-
-//------------------------------------------------------------------------------
-
-CVideoLayerBridgeDRMPRIME::CVideoLayerBridgeDRMPRIME(std::shared_ptr<CDRMUtils> drm)
-  : m_DRM(drm)
-{
-}
-
-CVideoLayerBridgeDRMPRIME::~CVideoLayerBridgeDRMPRIME()
-{
-  Release(m_prev_buffer);
-  Release(m_buffer);
-}
-
-void CVideoLayerBridgeDRMPRIME::Disable()
-{
-  // disable video plane
-  struct plane* plane = m_DRM->GetPrimaryPlane();
-  m_DRM->AddProperty(plane, "FB_ID", 0);
-  m_DRM->AddProperty(plane, "CRTC_ID", 0);
-}
-
-void CVideoLayerBridgeDRMPRIME::Acquire(CVideoBufferDRMPRIME* buffer)
-{
-  // release the buffer that is no longer presented on screen
-  Release(m_prev_buffer);
-
-  // release the buffer currently being presented next call
-  m_prev_buffer = m_buffer;
-
-  // reference count the buffer that is going to be presented on screen
-  m_buffer = buffer;
-  m_buffer->Acquire();
-}
-
-void CVideoLayerBridgeDRMPRIME::Release(CVideoBufferDRMPRIME* buffer)
-{
-  if (!buffer)
-    return;
-
-  Unmap(buffer);
-  buffer->Release();
-}
-
-bool CVideoLayerBridgeDRMPRIME::Map(CVideoBufferDRMPRIME* buffer)
-{
-  if (buffer->m_fb_id)
-    return true;
-
-  AVDRMFrameDescriptor* descriptor = buffer->GetDescriptor();
-  uint32_t handles[4] = {0}, pitches[4] = {0}, offsets[4] = {0}, flags = 0;
-  uint64_t modifier[4] = {0};
-  int ret;
-
-  // convert Prime FD to GEM handle
-  for (int object = 0; object < descriptor->nb_objects; object++)
-  {
-    ret = drmPrimeFDToHandle(m_DRM->GetFileDescriptor(), descriptor->objects[object].fd, &buffer->m_handles[object]);
-    if (ret < 0)
-    {
-      CLog::Log(LOGERROR, "CVideoLayerBridgeDRMPRIME::%s - failed to convert prime fd %d to gem handle %u, ret = %d",
-                __FUNCTION__, descriptor->objects[object].fd, buffer->m_handles[object], ret);
-      return false;
-    }
-  }
-
-  AVDRMLayerDescriptor* layer = &descriptor->layers[0];
-
-  for (int plane = 0; plane < layer->nb_planes; plane++)
-  {
-    int object = layer->planes[plane].object_index;
-    uint32_t handle = buffer->m_handles[object];
-    if (handle && layer->planes[plane].pitch)
-    {
-      handles[plane] = handle;
-      pitches[plane] = layer->planes[plane].pitch;
-      offsets[plane] = layer->planes[plane].offset;
-      modifier[plane] = descriptor->objects[object].format_modifier;
-    }
-  }
-
-  if (modifier[0] && modifier[0] != DRM_FORMAT_MOD_INVALID)
-    flags = DRM_MODE_FB_MODIFIERS;
-
-  // add the video frame FB
-  ret = drmModeAddFB2WithModifiers(m_DRM->GetFileDescriptor(), buffer->GetWidth(), buffer->GetHeight(), layer->format,
-                                   handles, pitches, offsets, modifier, &buffer->m_fb_id, flags);
-  if (ret < 0)
-  {
-    CLog::Log(LOGERROR, "CVideoLayerBridgeDRMPRIME::%s - failed to add fb %d, ret = %d", __FUNCTION__, buffer->m_fb_id, ret);
-    return false;
-  }
-
-  Acquire(buffer);
-  return true;
-}
-
-void CVideoLayerBridgeDRMPRIME::Unmap(CVideoBufferDRMPRIME* buffer)
-{
-  if (buffer->m_fb_id)
-  {
-    drmModeRmFB(m_DRM->GetFileDescriptor(), buffer->m_fb_id);
-    buffer->m_fb_id = 0;
-  }
-
-  for (int i = 0; i < AV_DRM_MAX_PLANES; i++)
-  {
-    if (buffer->m_handles[i])
-    {
-      struct drm_gem_close gem_close = { .handle = buffer->m_handles[i] };
-      drmIoctl(m_DRM->GetFileDescriptor(), DRM_IOCTL_GEM_CLOSE, &gem_close);
-      buffer->m_handles[i] = 0;
-    }
-  }
-}
-
-void CVideoLayerBridgeDRMPRIME::Configure(CVideoBufferDRMPRIME* buffer)
-{
-}
-
-void CVideoLayerBridgeDRMPRIME::SetVideoPlane(CVideoBufferDRMPRIME* buffer, const CRect& destRect)
-{
-  if (!Map(buffer))
-  {
-    Unmap(buffer);
-    return;
-  }
-
-  struct plane* plane = m_DRM->GetPrimaryPlane();
-  m_DRM->AddProperty(plane, "FB_ID", buffer->m_fb_id);
-  m_DRM->AddProperty(plane, "CRTC_ID", m_DRM->GetCrtc()->crtc->crtc_id);
-  m_DRM->AddProperty(plane, "SRC_X", 0);
-  m_DRM->AddProperty(plane, "SRC_Y", 0);
-  m_DRM->AddProperty(plane, "SRC_W", buffer->GetWidth() << 16);
-  m_DRM->AddProperty(plane, "SRC_H", buffer->GetHeight() << 16);
-  m_DRM->AddProperty(plane, "CRTC_X", static_cast<int32_t>(destRect.x1) & ~1);
-  m_DRM->AddProperty(plane, "CRTC_Y", static_cast<int32_t>(destRect.y1) & ~1);
-  m_DRM->AddProperty(plane, "CRTC_W", (static_cast<uint32_t>(destRect.Width()) + 1) & ~1);
-  m_DRM->AddProperty(plane, "CRTC_H", (static_cast<uint32_t>(destRect.Height()) + 1) & ~1);
 }
