+++ xbmc/cores/VideoPlayer/VideoRenderers/HwDecRender/RendererDRMPRIME.h
@@ -8,33 +8,10 @@
 
 #pragma once
 
-#include "cores/VideoPlayer/DVDCodecs/Video/DVDVideoCodecDRMPRIME.h"
 #include "cores/VideoPlayer/VideoRenderers/BaseRenderer.h"
-#include "windowing/gbm/WinSystemGbmEGLContext.h"
 
-class CVideoLayerBridgeDRMPRIME
-  : public CVideoLayerBridge
-{
-public:
-  CVideoLayerBridgeDRMPRIME(std::shared_ptr<CDRMUtils> drm);
-  ~CVideoLayerBridgeDRMPRIME();
-  void Disable() override;
-
-  virtual void Configure(CVideoBufferDRMPRIME* buffer);
-  virtual void SetVideoPlane(CVideoBufferDRMPRIME* buffer, const CRect& destRect);
-
-protected:
-  std::shared_ptr<CDRMUtils> m_DRM;
-
-private:
-  void Acquire(CVideoBufferDRMPRIME* buffer);
-  void Release(CVideoBufferDRMPRIME* buffer);
-  bool Map(CVideoBufferDRMPRIME* buffer);
-  void Unmap(CVideoBufferDRMPRIME* buffer);
-
-  CVideoBufferDRMPRIME* m_buffer = nullptr;
-  CVideoBufferDRMPRIME* m_prev_buffer = nullptr;
-};
+class CVideoBuffer;
+class CVideoLayerBridgeDRMPRIME;
 
 class CRendererDRMPRIME
   : public CBaseRenderer
